requirements.txt
gradio
requests
beautifulsoup4
sentence-transformers
scikit-learn
python-dotenv
openai